Move unit tests around into per-platform where appropriate.
